Add Ability subclasses
Let's move the AbilityManager into the Ability subdirectory and add the following Ability subclasses:
- BasicProjectileAbility
- ExplosiveProjectileAbility
- LockedOnAbility
- LaserAbility
- ZonalAoeAbility
- AuraAbility
- OrbitalAbility
- ConalAbility

Also, add methods to Ability and AbilityManager to store a reference to the Player GameObject, from which the transform is obtained. I don't think this is the best way to go about this, will try to improve this later.